fix: repository 수정

+++ server/src/main/java/com/seniclass/server/domain/category/repository/SubCategoryRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubCategoryRepository extends JpaRepository<SubCategory, Integer> {}
+public interface SubCategoryRepository extends JpaRepository<SubCategory, Long> {}
